Refatoração da estrutura de telemetria e integração SignalR
Atualiza a estrutura `SPageFilePhysics` com novos campos e comentários.
Modifica a classe `TelemetryData` para o namespace correto e altera tipos de propriedades.
Define a interface `ITelemetryRepository` e ajusta o repositório para o novo namespace.
Configura o `Program.cs` para incluir serviços de telemetria e CORS, além de mapear o hub do SignalR.
Adiciona o `GetTelemetryHub` e implementa o `TelemetryWorker` para enviar dados de telemetria em tempo real.

+++ ACRH/Entities/SPageFilePhysics.cs
@@ -7,10 +7,12 @@ public unsafe struct SPageFilePhysics
     public float Gas;
     public float Brake;
     public float Fuel;
-    public float Gear;
-    public float Rpms;
+    public int Gear;
+    public int Rpms;
     public float SteerAngle;
     public float SpeedKmh;
+    //aqui ta errado, tem mts campos entre speedkmh e os embaixo,
+    //lembrar de aadicionar ou pular eles depois
     public fixed float WheelAngularSpeed[4];
     public fixed float TyreWear[4];
     public fixed float TyreTemp[4];
